refactor: clean up grpc context exports
