--- packages/insomnia-app/app/ui/components/panes/use-grpc-ipc.js
@@ -1,9 +1,0 @@
-// @flow
-import React from 'react';
-import type { GrpcRequestEvent } from '../../../common/grpc-events';
-import { ipcRenderer } from 'electron';
-
-export const useGrpcIpc: string => GrpcRequestEvent => void = requestId =>
-  React.useCallback((channel: GrpcRequestEvent) => ipcRenderer.send(channel, requestId), [
-    requestId,
-  ]);
